Mandatory limit on "sigma"
It always took a long time (more than 24h) to run the simulation when considering the combination of TP and FH. This might be caused by a significant reduction of effective normal stress. So a mandatory limit has been added here (sigma = 1MPa when sigma < 1MPa). Performance of the code has been improved but reasonability still need to be checked through modeling under different conditions.
Bowen, 2023/6/1